Updating in accordance to updates in submodule, as well as trying to update leaphand node
