+++ config/test_params.yaml
@@ -7,6 +7,6 @@ pub_current: false
 pub_pos: true
 pub_vel: false
 start_pos: [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-joint_feedback_topic: /dom/dynamixel_joint_states
-joint_command_topic: /dom/command_joint_states
+joint_feedback_topic: /dom/end_eff/dynamixel_joint_states
+joint_command_topic: /dom/end_eff/command_joint_states
 dynamixel_type: "XL330-M288"
