+++ package.xml
@@ -22,6 +22,9 @@
     <exec_depend>python3-yaml</exec_depend>
     <exec_depend>ros2launch</exec_depend>
 
+    <!--installing my python package-->
+    <buildtool_depend>ament_cmake_python</buildtool_depend>
+
     <!--unit tests -->
     <test_depend>ament_cmake_pytest</test_depend>
     <test_depend>pytest</test_depend>
